+++ src/renderer/components/MarkdownEditor.vue
@@ -12,9 +12,7 @@
             </b-alert>
         </div>
 
-        <loading :active.sync="isLoading"
-                 :can-cancel="false"
-                 :is-full-page="true"></loading>
+        <loading :active="isLoading"></loading>
 
         <div class="h-100 d-flex bd-highlight align-content-stretch flex-column">
 
@@ -102,12 +100,14 @@
 
 <script>
     import {ipcRenderer} from 'electron'; // eslint-disable-line
-    import Loading from 'vue-loading-overlay';
+    // import Loading from 'vue-loading-overlay';
 
     import Vue from 'vue';
 
     import marked from 'marked';
     import highlight from 'highlight.js';
+
+    import Loading from './Loading';
 
     import {
       SHOW_OPEN_DIALOG,
@@ -121,7 +121,7 @@
 
     const ace = require('brace');
     require('brace/mode/markdown');
-    require('brace/theme/monokai');
+    require('brace/theme/terminal');
 
     const bus = new Vue();
 
@@ -256,7 +256,7 @@
 
         this.aceEditor = ace.edit(editor);
         this.aceEditor.getSession().setMode('ace/mode/markdown');
-        this.aceEditor.setTheme('ace/theme/monokai');
+        this.aceEditor.setTheme('ace/theme/terminal');
         this.aceEditor.getSession().setUseWrapMode(true);
         this.aceEditor.getSession().setWrapLimitRange(80, 100);
         this.aceEditor.$blockScrolling = Infinity;